requirements.txt
django
django-dotenv
django-environ
whitenoise
django-crispy-forms
crispy-bootstrap4
Pillow
django-phonenumber-field[phonenumbers]
